Separate Solution tests from Project tests
